Lib

Documentation

data Planet

Enum with all the Planet types

Constructors

Ferengi

Betasoide

Vulcano

data Point

Representations of a point in space

Constructors

Point

Fields

x :: Distance

y :: Distance

type Angle = Double

Dummy type for angles

type Day = Double

Dummy type for days

type Speed = Double

Dummy type for speeds

type Distance = Double

Dummy type for distances in space

type Distance' = Double

Dummy type for derivated of distance in space

data Line

A line is described as two points

Constructors

Line

Fields

from :: Point

to :: Point

data Side

Side is useful to tell to which side of a line is a point

Constructors

Above

Below

data Slope

Slope is useful to tell if a distance is growing or shrinking

Constructors

Descending

Flat

Ascending

data Weather

The weather of the planets which depends on their relative positions

Constructors

Drought

HeavyRain

Rain

Perfect

Unknown

weatherAt :: Day -> Weather

Entry point for most programs which checks the weather conditions for a given day

sideOfOtherPlanet :: Day -> Maybe Side

pmod' :: (Num a, Ord a) => a -> a -> a

pmod :: Angle -> Angle -> Angle

isSunInsideTriangle :: Angle -> Angle -> Angle -> Bool

Using the angles check if the sun is inside the triangle if for each planet the other two planets are on different sides of the sun then the sun must be inside the triangle that all planets form

isSunAlignedWithPlanets :: Angle -> Angle -> Angle -> Bool

Check if the angles of planets are all aligned with the sun If mod 180 all angles are the same then all planets are aligned

isSunInsideTriangleAt :: Day -> Bool

isSunAlignedWithPlanetsAt :: Day -> Bool

sideOf :: Point -> Line -> Maybe Side

Check to which side of a line a point is

If the point is on the line then return Nothing

distanceBetween :: Point -> Point -> Distance

allPlanets :: [Planet]

otherPlanets :: Planet -> [Planet]

distanceBetweenPlanets :: Day -> Planet -> Planet -> Distance

distanceBetweenPlanets' :: Day -> Planet -> Planet -> Distance'

distanceBetweenAllPlanets :: Day -> Distance

slopeOfTriangleAt :: Day -> Slope

distanceBetweenAllPlanets' :: Day -> Distance

distanceBetweenPlanets2 :: Day -> Planet -> Planet -> Distance

toRadians :: Angle -> Angle

toDegrees :: Angle -> Angle

positionAfterDays :: Day -> Planet -> Point

Get the angle of a planet at a given day

angleAfterDays :: Day -> Planet -> Angle

Get the angle of a planet at a given day

distanceFromTheSun :: Planet -> Distance

Get the distance from the Sun of a given planet

speed :: Planet -> Speed

Get the speed of a given planet